pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["ksredkin <ksredkin0968@gmail.com>"]
readme = "docs/README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.124.4"
uvicorn = "^0.38.0"
python-jose = "^3.5.0"
passlib = "^1.7.4"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
